Fix(YoloSimulation): Fix missing null check in stopping
Stop should be safe to call at any time. Therefore make sure that we
handle uninitialized states, where e.g. the streams could still be null as
well.

+++ server/processes/YoloSimulation.js
@@ -140,8 +140,12 @@ const start = function () {
 
 const stop = function () {
   return new Promise((resolve, reject) => {
-    yoloSimulation.simulationMJPEGServer.kill();
-    yoloSimulation.simulationJSONHTTPStreamServer.kill();
+    if(yoloSimulation.simulationMJPEGServer) {
+      yoloSimulation.simulationMJPEGServer.kill();
+    }
+    if(yoloSimulation.simulationJSONHTTPStreamServer) {
+      yoloSimulation.simulationJSONHTTPStreamServer.kill();
+    }
     yoloSimulation.isStarted = false;
     resolve();
   });
